Begin splitting Payment Summary pages
