+++ app/controllers/payments_controller.rb
@@ -62,6 +62,7 @@ class PaymentsController < ApplicationController
   def summary
     add_payment_summary_breadcrumb
     authorize Payment.new
+    @expense_groups = ExpenseGroup.includes(:translations)
     @expense_items = ExpenseItem.includes(:translations, expense_group: [:translations]).ordered
   end
