+++ app/views/payment_summary/expense_groups/index.html.haml
@@ -1,0 +1,1 @@
+index.html.haml
